requirements.txt
certifi==2020.6.20
chardet==3.0.4
future==0.15.2
gspread==0.4.1
httplib2==0.9.2
idna==2.10
mxnet-cu92==1.5.0
oauth2client==2.2.0
oauthlib==3.1.0
pprintpp==0.2.3
py==1.4.31
pyasn1==0.4.8
pyasn1-modules==0.2.8
PySocks==1.7.1
pytest==2.9.2
requests-oauthlib==1.3.0
rsa==3.4.2
selenium==3.141.0
six==1.10.0
tweepy==3.9.0
urllib3==1.25.11
wincertstore==0.2
xmltodict==0.12.0
